[YS-151] refactor: 학교 메일 인증 컴포넌트 분리
