+++ src/app/join/Join.types.ts
@@ -1,0 +1,10 @@
+export interface FormInput {
+  socialEmail: string;
+  contactEmail: string;
+  univEmail: string;
+  authCode: string;
+  isAllCheck: boolean;
+  isTermOfService: boolean;
+  isPrivacy: boolean;
+  isAdvertise: boolean;
+}
